fix(auth): use areTokensValid instead of hasAuthenticated
Comment from Qui:

`hasAuthenticated` indicates the user has authenticated at least once.
This is true even when the refresh token expires.
I'll add documentation later to make things clearer.
Change this to use the areTokensValid property.
This means the refresh token expiration is in the future and the user won't need to re-authenticate any time soon.
So when areTokensValid = true the login options don't need to be presented.
When false the user will want to re-authenticate to get a new refresh token.